Add new component for fullscreen mode [MAILPOET-3205]
